requirements.txt
twitch-python
RPi.GPIO
pillow
aiortc
